backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
corsheaders==3.14.0
requests==2.31.0
Pillow==10.0.0
python-dotenv